Move Scorekeeper into its own component and display above fishbowl during game play
