+++ ui/src/components/ScoreKeeper.css
@@ -1,0 +1,34 @@
+.scorekeeper {
+    margin-top: 5px;
+    border-radius: 3px;
+    height: 100px;
+    min-width: 300px;
+    width: 800px;
+  }
+  
+  .score {
+    margin-left: 5%;
+    margin-right: 5%;
+    height: 90%;
+    text-align: center;
+  }
+
+  .score > .col-left {
+    float: left;
+    width: 50%;
+  }
+  
+  .score > col-right {
+    margin-left: 50%;
+  }
+
+  @media screen and (max-width: 600px) {
+    .col-left {
+      width: 100%;
+      text-align: center;
+    }
+    .col-right {
+      margin-left: 0;
+      text-align: center;
+    }
+  }
